Switching over to use GitPython
