+++ app/pip/2.7/app/requirements.txt
@@ -1,2 +1,2 @@
 docopt
-plumbum
+GitPython
