+++ app/pip/3.5/app/requirements.txt
@@ -1,2 +1,2 @@
 docopt
-plumbum
+GitPython
